Move to Ocean submodule
- Move PALEOocean -> PALEOocean.Ocean
- Update documentation
- Update to v0.2.0
