+++ src/PALEOocean.jl
@@ -9,8 +9,7 @@ function moduledir()
     return dirname(@__DIR__)
 end
 
-include("OceanBase.jl")
+include("ocean/Ocean.jl")
 
-include("OceanNoTransport.jl")
 
 end # module PALEOocean
